+++ CHANGELOG.md
@@ -7,10 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Radar model, RadarURLParser, RadarURLBuilder protocols, implementation, and tests
-- RadarOption and BrowserOption
 - Basic UI architecture
 - BriskRadar and tests
 - Localization
 - Present in SFSafariViewController
-- Ask open in which browser
 - Use Main.storyboard as launch screen file
+- RadarOption and BrowserOption in settings